# Use system-wide thread ID in info log lines